+++ AIQuotaMonitor/Domain/ProviderContracts.swift
@@ -10,10 +10,26 @@ struct ProviderFetchResult: Sendable {
     let snapshot: ProviderSnapshot
 }
 
+enum ProviderRefreshPolicy: Sendable, Equatable {
+    case scheduled
+    case userInitiated
+
+    var allowsCachedSuccess: Bool {
+        self == .scheduled
+    }
+}
+
 protocol QuotaProvider: Sendable {
     var id: ProviderID { get }
     func availability() async -> ProviderAvailability
     func fetchQuota() async -> ProviderFetchResult
+    func fetchQuota(policy: ProviderRefreshPolicy) async -> ProviderFetchResult
+}
+
+extension QuotaProvider {
+    func fetchQuota(policy: ProviderRefreshPolicy) async -> ProviderFetchResult {
+        await fetchQuota()
+    }
 }
 
 protocol LocalUsageSource: Sendable {